+++ sdk/keyvault/azure-security-keyvault-administration/inc/azure/keyvault/administration/rest_client_models.hpp
@@ -74,4 +74,191 @@ namespace Azure { namespace Security { namespace KeyVault { namespace Administra
       std::vector<Setting> Value;
     };
 
+    /**
+     * @brief Define the options to create an SDK Keys client.
+     *
+     */
+    struct BackupRestoreClientOptions final : public Azure::Core::_internal::ClientOptions
+    {
+      /**
+       * @brief Service Version used.
+       *
+       */
+      const std::string ApiVersion{"7.5"};
+    };
+
+    /**
+     * @brief KeyVault Service Error model.
+     *
+     */
+    struct KeyVaultServiceError final
+    {
+      /**
+       * @brief The error code.
+       *
+       */
+      std::string Code;
+      /**
+       * @brief The error message.
+       *
+       */
+      std::string Message;
+      /**
+       * @brief Inner error details.
+       *
+       */
+      std::unique_ptr<KeyVaultServiceError> InnerError;
+    };
+
+    /**
+     * @brief The full backup operation.
+     *
+     */
+    struct FullBackupOperation final
+    {
+      /**
+       * @brief Status of the backup operation.
+       *
+       */
+      std::string Status;
+      /**
+       * @brief The status details of backup operation.
+       *
+       */
+      Azure::Nullable<std::string> StatusDetails;
+      /**
+       * @brief Error encountered, if any, during the full backup operation.
+       *
+       */
+      Azure::Nullable<KeyVaultServiceError> Error;
+      /**
+       * @brief The start time of the backup operation in UTC.
+       *
+       */
+      DateTime StartTime;
+      /**
+       * @brief The end time of the backup operation in UTC.
+       *
+       */
+      Nullable<DateTime> EndTime;
+      /**
+       * @brief Identifier for the full backup operation.
+       *
+       */
+      std::string JobId;
+      /**
+       * @brief The Azure blob storage container Uri which contains the full backup.
+       *
+       */
+      std::string AzureStorageBlobContainerUri;
+    };
+
+    /**
+     * @brief Sas token parameter for backup and restore operations.
+     *
+     */
+    struct SasTokenParameter final
+    {
+      /**
+       * @brief The SAS token pointing to an Azure Blob storage container.
+       *
+       */
+      Nullable<std::string> Token;
+      /**
+       * @brief Indicates which authentication method should be used. If set to true, Managed HSM
+       * will use the configured user-assigned managed identity to authenticate with Azure Storage.
+       * Otherwise, a SAS token has to be specified.
+       *
+       */
+      Nullable<bool> UseManagedIdentity;
+    };
+
+    /**
+     * @brief Full backup status options.
+     *
+     */
+    struct FullBackupStatusOptions final
+    {
+      /**
+       * @brief Identifier for the full backup operation.
+       *
+       */
+      std::string JobId;
+    };
+
+    /**
+     * @brief Full restore operation.
+     *
+     */
+    struct RestoreOperation final
+    {
+      /**
+       * @brief Status of the restore operation.
+       *
+       */
+      std::string Status;
+      /**
+       * @brief The status details of restore operation.
+       *
+       */
+      Azure::Nullable<std::string> StatusDetails;
+      /**
+       * @brief Error encountered, if any, during the restore operation.
+       *
+       */
+      Azure::Nullable<KeyVaultServiceError> Error;
+      /**
+       * @brief Identifier for the restore operation.
+       *
+       */
+      std::string JobId;
+      /**
+       * @brief The start time of the restore operation.
+       *
+       */
+      DateTime StartTime;
+      /**
+       * @brief The end time of the restore operation.
+       *
+       */
+      Nullable<DateTime> EndTime;
+    };
+
+    /**
+     * @brief The selective key restore operation.
+     *
+     */
+    struct SelectiveKeyRestoreOperation final
+    {
+      /**
+       * @brief The status of the selective key restore operation.
+       *
+       */
+      std::string Status;
+      /**
+       * @brief The status details of the selective key restore operation.
+       *
+       */
+      Azure::Nullable<std::string> StatusDetails;
+      /**
+       * @brief Error encountered, if any, during the selective key restore operation.
+       *
+       */
+      Azure::Nullable<KeyVaultServiceError> Error;
+      /**
+       * @brief Identifier for the selective key restore operation.
+       *
+       */
+      std::string JobId;
+      /**
+       * @brief The start time of the selective key restore operation.
+       *
+       */
+      DateTime StartTime;
+      /**
+       * @brief The end time of the selective key restore operation.
+       *
+       */
+      Nullable<DateTime> EndTime;
+    };
 }}}}} // namespace Azure::Security::KeyVault::Administration::Models